refactor(presence): fold agent-core-presence into agent-core-webcam
