--- packages/agent-core-presence/pyproject.toml
@@ -1,37 +1,0 @@
-[project]
-name = "agent-core-presence"
-dynamic = ["version"]
-description = "Presence/identity injector hook for agent_core — reads a local presence-state file and injects a compact, staleness-guarded presence tag into session context"
-license = "MIT"
-requires-python = ">=3.12"
-dependencies = [
-    "agent-core-bus>=0.8,<0.9",
-    "pluggy>=1.6",
-    "pydantic>=2.7",
-]
-
-[project.optional-dependencies]
-# The continuous CV watcher (Phase 2) — kept optional so the hook install stays
-# tiny. opencv + a face-detection model get added here once approved.
-cv = [
-    "numpy>=1.26",
-]
-
-[project.entry-points."agent_core"]
-presence_aliases = "agent_core_presence.plugin"
-
-[build-system]
-requires = ["hatchling", "uv-dynamic-versioning"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/agent_core_presence"]
-
-[tool.hatch.version]
-source = "uv-dynamic-versioning"
-
-[tool.uv-dynamic-versioning]
-fallback-version = "0.0.0"
-
-[tool.uv]
-cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true, tags = true } }]
