+++ packages/agent-core-webcam/src/agent_core_webcam/presence/motion.py
